The last time I tried this the monkey didn't survive. Let's hope it works better this time.

+++ main_test.go
@@ -2,10 +2,11 @@ package main
 
 import "testing"
 
+// Without inlining BenchmarkSimpleBitmapIndex-12    	  120832	      9747 ns/op	       0 B/op	       0 allocs/op
 func BenchmarkSimpleBitmapIndex(b *testing.B) {
 	_, _, terrace, reservations, _, expensive := initData()
 
-	resBitmap := make([]byte, bitmapLength)
+	resBitmap := make([]uint64, bitmapLength)
 
 	b.ResetTimer()
 	b.ReportAllocs()
